+++ src/registry.ts
@@ -8,7 +8,9 @@ import * as fss from "fs";
 
 import * as fileutil from "./fileutil";
 import logger from "./logger";
-import { Config, Image, Layer, Manifest, PartialManifestConfig } from "./types";
+import {Config, Image, Index, Layer, Manifest, IndexManifest, PartialManifestConfig, Platform} from "./types";
+import {DockerV2, OCI} from "./MIMETypes";
+import {getLayerTypeFileEnding} from "./utils";
 
 type Headers = Record<string, string>;
 
@@ -160,7 +162,7 @@ export function createRegistry(registryBaseUrl: string, token: string) {
 
 	async function uploadLayerContent(uploadUrl: string, layer: Layer, dir: string) {
 		logger.info(layer.digest);
-		const file = path.join(dir, getHash(layer.digest) + (layer.mediaType.includes("tar.gzip") ? ".tar.gz" : ".tar"));
+		const file = path.join(dir, getHash(layer.digest) + getLayerTypeFileEnding(layer));
 		await uploadContent(uploadUrl, file, layer, auth);
 	}
 
@@ -190,20 +192,64 @@ export function createRegistry(registryBaseUrl: string, token: string) {
 		});
 	}
 
-	async function dlManifest(image: Image): Promise<Manifest> {
-		return await dlJson(
+	async function dlManifest(image: Image, preferredPlatform: Platform): Promise<Manifest> {
+
+		// Accept both manifests and index/manifest lists
+		const res = await dlJson<Manifest|Index>(
 			`${registryBaseUrl}${image.path}/manifests/${image.tag}`,
-			buildHeaders("application/vnd.docker.distribution.manifest.v2+json", auth),
+			buildHeaders(`${OCI.index}, ${OCI.manifest}, ${DockerV2.index}, ${DockerV2.manifest}`, auth),
 		);
+
+		// We've received an OCI Index or Docker Manifest List and need to find which manifest we want
+		if (res.mediaType === OCI.index || res.mediaType === DockerV2.index) {
+			const availableManifests = (res as Index).manifests
+			const adequateManifest = pickManifest(availableManifests, preferredPlatform)
+			return dlManifest({...image, tag: adequateManifest.digest}, preferredPlatform)
+		}
+
+		return res as Manifest
+	}
+
+	function pickManifest(manifests: IndexManifest[], preferredPlatform: Platform): IndexManifest {
+		const matchingArchitectures = new Set<IndexManifest>;
+		const matchingOSes = new Set<IndexManifest>;
+		// Find sets of matching architecture and os
+		for (const manifest of manifests) {
+			if (manifest.platform.architecture === preferredPlatform.architecture) {
+				matchingArchitectures.add(manifest)
+			}
+			if (manifest.platform.os === preferredPlatform.os) {
+				matchingOSes.add(manifest)
+			}
+		}
+
+		// If the intersection of matching architectures and OS is one we've found our optimal match
+		const intersection = new Set([...matchingArchitectures].filter((x) => matchingOSes.has(x)));
+		if (intersection.size == 1) {
+			return intersection.values().next().value
+		}
+
+		// If we don't have a perfect match we give a warning and try the first matching architecture
+		if (matchingArchitectures.size >= 1) {
+			const matchingArch = matchingArchitectures.values().next().value
+			logger.info(`[WARN] Preferred OS '${preferredPlatform.os}' not available.`)
+			logger.info('[WARN] Using closest available manifest:', JSON.stringify(matchingArch.platform))
+			return matchingArch
+		}
+
+		// If there's no image matching the wanted architecture we bail
+		logger.error(`No image matching requested architecture: '${preferredPlatform.architecture}'`)
+		logger.error('Available platforms:', JSON.stringify(manifests.map(m => m.platform)))
+		throw new Error('No image matching requested architecture')
 	}
 
 	async function dlConfig(image: Image, config: Manifest["config"]): Promise<Config> {
-		return await dlJson(`${registryBaseUrl}${image.path}/blobs/${config.digest}`, buildHeaders("*/*", auth));
+		return await dlJson<Config>(`${registryBaseUrl}${image.path}/blobs/${config.digest}`, buildHeaders("*/*", auth));
 	}
 
 	async function dlLayer(image: Image, layer: Layer, folder: string): Promise<string> {
 		logger.info(layer.digest);
-		const file = getHash(layer.digest) + (layer.mediaType.includes("tar.gzip") ? ".tar.gz" : ".tar");
+		const file = getHash(layer.digest) + getLayerTypeFileEnding(layer);
 		await dlToFile(
 			`${registryBaseUrl}${image.path}/blobs/${layer.digest}`,
 			path.join(folder, file),
@@ -252,15 +298,23 @@ export function createRegistry(registryBaseUrl: string, token: string) {
 		);
 	}
 
-	async function download(imageStr: string, folder: string) {
+	async function download(imageStr: string, folder: string, preferredPlatform: Platform) {
 		const image = parseImage(imageStr);
 
 		logger.info("Downloading manifest...");
-		const manifest = await dlManifest(image);
+		const manifest = await dlManifest(image, preferredPlatform);
 		await fs.writeFile(path.join(folder, "manifest.json"), JSON.stringify(manifest));
 
 		logger.info("Downloading config...");
 		const config = await dlConfig(image, manifest.config);
+
+		if (config.architecture != preferredPlatform.architecture) {
+			logger.info(`[WARN] Image architecture (${config.architecture}) does not match preferred architecture (${preferredPlatform.architecture}).`)
+		}
+		if (config.os != preferredPlatform.os) {
+			logger.info(`[WARN] Image OS (${config.os}) does not match preferred OS (${preferredPlatform.os}).`)
+		}
+
 		await fs.writeFile(path.join(folder, "config.json"), JSON.stringify(config));
 
 		logger.info("Downloading layers...");
@@ -285,10 +339,10 @@ export function createDockerRegistry(auth?: string) {
 		);
 		return resp.token;
 	}
-	async function download(imageStr: string, folder: string) {
+	async function download(imageStr: string, folder: string, platform: Platform) {
 		const image = parseImage(imageStr);
 		if (!auth) auth = await getToken(image);
-		await createRegistry(registryBaseUrl, auth).download(imageStr, folder);
+		await createRegistry(registryBaseUrl, auth).download(imageStr, folder, platform);
 	}
 
 	async function upload(imageStr: string, folder: string) {