+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.1.0] - 2023-03-04
+
+### Added
+
+- Support for OCI Image Index (application/vnd.oci.image.index.v1)
+- Support for OCI Image Manifest (application/vnd.oci.image.manifest.v1)
+- Support for Docker Manifest List (application/vnd.docker.distribution.manifest.list.v2)
+- Option for selecting preferred platform (OS and architecture) similar to Docker 
+
+### Improvement
+
+- Add `doqr` version in manifest history `created_by` field
+
 ## [1.0.4] - 2023-02-07
 
 ### Fixed